+++ src/curves/math/spline.hpp
@@ -27,7 +27,7 @@ static_assert(SPLINE_SEGMENTS_PER_OCTAVE_LOG2 <=
               SPLINE_OCTAVE_ORIGIN_FIXED_SHIFT);
 
 // Imports from c.
-auto map_x_to_t(s64 x, int width_log2) noexcept -> s64;
+auto map_x_to_t(s64 x, unsigned int width_log2) noexcept -> s64;
 auto map_v_to_x(const struct curves_spline* spline, s64 v) noexcept -> s64;
 auto calc_segment_desc(s64 x) noexcept -> curves_segment_desc;
 auto resolve_x(s64 x) noexcept -> curves_spline_coords;
@@ -309,8 +309,8 @@ class SplineBuilder {
     const auto y_start = prev_a + prev_b + prev_c + prev_d;
 
     // Calculate normalized derivatives at the boundary.
-    const auto m_start_norm = 3.0 * prev_a + 2.0 * prev_b + prev_c;
-    const auto k_start_norm = 6.0 * prev_a + 2.0 * prev_b;
+    const auto m_start_norm = 3.0L * prev_a + 2.0L * prev_b + prev_c;
+    const auto k_start_norm = 6.0L * prev_a + 2.0L * prev_b;
 
     // Get previous physical width (dv_prev) to un-normalize.
     const s64 prev_len_fixed = 1ULL << prev_width_log2;
@@ -332,19 +332,19 @@ class SplineBuilder {
     real_t dv_final = 0;
 
     // Safety floor nominally 0.
-    const real_t kMinSlope = 0.0;
+    const real_t kMinSlope = 0.0L;
 
     // Adaptive loop.
     //
     // Search for the widest smooth segment that doesn't dive.
     for (; runout_log2 >= SPLINE_MIN_SEGMENT_WIDTH_LOG2; --runout_log2) {
       // Calculate candidate width.
-      const s64 runout_len_fixed = 1ULL << runout_log2;
+      const s64 runout_len_fixed = 1LL << runout_log2;
       dv_final = Fixed::from_raw(runout_len_fixed).to_real() * x_to_v;
 
       // Predict end slope (assuming standard y''(1)=0 constraint)
       // slope_end = slope_start + 0.5*curvature_start*width
-      const real_t projected_slope = m_real + 0.5 * k_real * dv_final;
+      const real_t projected_slope = m_real + 0.5L * k_real * dv_final;
 
       if (projected_slope >= kMinSlope) {
         is_safe = true;
@@ -360,13 +360,13 @@ class SplineBuilder {
     if (!is_safe) {
       // Force use of minimum width
       runout_log2 = SPLINE_MIN_SEGMENT_WIDTH_LOG2;
-      const s64 min_len = 1ULL << runout_log2;
+      const s64 min_len = 1LL << runout_log2;
       dv_final = Fixed::from_raw(min_len).to_real() * x_to_v;
     }
 
     // Commit runout length.
-    result.runout_width_log2 = runout_log2;
-    result.x_runout_limit = result.x_geometric_limit + (1ULL << runout_log2);
+    result.runout_width_log2 = (unsigned int)runout_log2;
+    result.x_runout_limit = result.x_geometric_limit + (1LL << runout_log2);
 
     // Calc constant segment coefficients.
     //
@@ -374,7 +374,7 @@ class SplineBuilder {
     // to the final chosen width, dv_final.
     const auto next_d = y_start;
     const auto next_c = m_real * dv_final;
-    const auto next_b = (k_real * dv_final * dv_final) / 2.0;
+    const auto next_b = (k_real * dv_final * dv_final) / 2.0L;
 
     // Calc cubic segment coefficient.
     real_t next_a;
@@ -382,13 +382,13 @@ class SplineBuilder {
       // We found a safe width that pulls out of the dive and ends with zero
       // curvature (y'' = 0). This is a smooth bleed-off and it maintains C2
       // continuity.
-      next_a = -next_b / 3.0;
+      next_a = -next_b / 3.0L;
     } else {
       // [PANIC MODE]
       // We did not find a safe width that pulls out ofthe dive.
       // Force the curve to flatten. Breaks C2 with a jerk spike, but
       // preserves monotonicity. Derived from 3a + 2b + c = 0, when y' = 0.
-      next_a = -(2.0 * next_b + next_c) / 3.0;
+      next_a = -(2.0L * next_b + next_c) / 3.0L;
     }
 
     result.runout_segment.coeffs[0] = Fixed{next_a}.raw;
